Add Windows build support (v1.2.0)
Cross-platform Bluetooth handling:
- bluetooth.go: Shared types and errors
- bluetooth_linux.go: Linux-specific rfcomm/bluetoothctl implementation
- bluetooth_windows.go: Windows COM port enumeration via registry

Build system:
- make build-windows: Cross-compile for Windows from Linux
- make build-all: Build both platforms
- make install-cross-deps: Install mingw-w64 for cross-compilation

On Windows, paired Bluetooth SPP devices appear as COM ports
automatically - no special rfcomm setup needed.

Updated README with:
- Platform support table
- Windows-specific instructions
- Cross-compilation guide
- Windows troubleshooting tips
